Creating an OMR Test grader for multiple choice bubble sheets.
Using a variety of Computer Vision techniques to build the solver by following an image processing pipeline.
Blurring and Edge detecting the image and then finding contours.
Applying a 4 point transform to achieve a top down image of the sheet of paper.
Sorting the contours from top to bottom and left to right for each question.
Then masking the image with contours for each bubble and determining the bubbled in answer to determine whether it is correct or not