Day 5 of intro to OpenCV. Creating a program to follow objects of a specified colour.
Using VideoStream or VideoCapture to either capture a videostream or operate on a given video file.
Resizing, blurring, and converting the frame to HSV to operate upon it.
Creating a mask for objects of the specified images and eroding and dilating it to remove small blobs.
Finding contours based off the mask and determining the largest contour as well as it's radius and center (x,y) coordinates.
Using a double ended queue with a maxsize to keep a moving contrail of the object with varying size depending on its distance from the object.